Добавил проверку версии wie; урезал лог; + удобный вывод.

+++ wine_pain_2.sh
@@ -4,7 +4,7 @@ LANG=C
 user1=$(who | grep '(:0)' | cut -d " " -f1)
 dir1=$(ls /home/$user1/.wine/drive_c/users/$user1/| wc -l)
 
-#Удаляем неудачные обновы
+#Удаляем неудачные бекапы
 rm -rf /.*[0-9]*
 
 #Чек размера диска и log cups чистка 
@@ -29,20 +29,22 @@ check5=/home/$user1/.wine/drive_c/FssTools
 check6=/home/$user1/.wine/drive_c/Radiant
 
 #Бекап бутылки
+result_message1=""
+
 if [ -d $check2 ] || [ -d $check3 ] || [ -d $check4 ] || [ -d $check5 ] || [ -d $check6 ]; then
   if [ -d .wine.special ]; then
     i=1
     while [ -d .wine.special$i ]; do
       i=$((i+1))
     done
-        echo "n-копия чего-то там"
+    result_message1="n-копия чего-то там"
     su - "$user1" -c "wineserver -k; sleep 3; mv .wine .wine.special$i; cp -r /etc/skel/.wine ."
   else
-      	echo "1 копия чего-то там"
+    result_message1="1 копия чего-то там"
     su - "$user1" -c "wineserver -k; sleep 3; mv .wine .wine.special; cp -r /etc/skel/.wine ."
   fi
 else
-    	echo "Приблуды не найдены"
+  result_message1="Приблуды не найдены"
   su - "$user1" -c "wineserver -k; sleep 3; rm -rf .wine.bak; mv .wine .wine.bak; cp -r /etc/skel/.wine ."
 fi
 
@@ -71,34 +73,45 @@ fi
 
 #Правильный запуск
 sed -i "5c\Exec=/usr/bin/run_vita.sh" /etc/skel/Рабочий\ стол/as_rmiac.desktop
-##! Не работает!!!
+
+
 # Проверяем бутылку и версию wine 
-check8=/etc/skel/.wine/wrapper.cfg
+result_message2=""
+#check8=/etc/skel/.wine/wrapper.cfg
 check9=/root/etersoft-repo/wine_bottle_8.0.6.tar.gz
-check10=/usr/bin/import_certs.sh
+#check10=/usr/bin/import_certs.sh
 
-########Проверяем бутылку и конфиг import_certs.sh#
-#if [ "$wine" == "wine-etersoft-8.0.5-alt0.M80P.1" ]; then
 
+########Проверяем бутылку
+if [ "$wine" == "wine-etersoft-8.0.5-alt0.M80P.1" ]; then
+	if [ test -e "$check9" ]; then 
+	echo
+	else
+		cd /root/etersoft-repo/
+	   wget http://10.11.128.115/.pcstuff/wine/wine_etersoft_repo_8.0.6/wine_bottle_8.0.6.tar.gz
+    	cd /etc/skel/
+		rm -rf .wine
+		tar -xvf /root/etersoft-repo/wine_bottle_8.0.6.tar.gz
+		result_message2="$(wine --version); Не было wine_bottle_8.0.6 - поставил." 
+	fi
+	else
+	  result_message2=$(wine --version)
+fi
 #if grep -q "wine /usr/lib/wine/i386-unix/cpcsp_proxy_setup.exe.so one1 two1 three1" "$check10"; then
- #   sed -i 's|wine /usr/lib/wine/i386-unix/cpcsp_proxy_setup.exe.so one1 two1 three1|wine /usr/lib/wine/cpcsp_proxy_setup.exe.so one1 two1|g' "$check10"
+#   sed -i 's|wine /usr/lib/wine/i386-unix/cpcsp_proxy_setup.exe.so one1 two1 three1|wine /usr/lib/wine/cpcsp_proxy_setup.exe.so one1 two1|g' "$check10"
 #else
- #   echo
+#   echo
 #fi
 
-    if test -e "$check8"; then
-			echo
-    else
-    if test -e "$check9"; then
-	echo
-else
-	cd /root/etersoft-repo/
-	wget http://10.11.128.115/.pcstuff/wine/wine_etersoft_repo_8.0.6/wine_bottle_8.0.6.tar.gz
-    		cd /etc/skel/
-			rm -rf .wine
-			tar -xvf /root/etersoft-repo/wine_bottle_8.0.6.tar.gz
-	  fi
-	fi
+#   if test -e "$check8"; then
+#			echo
+#   else
+#     if test -e "$check9"; then
+#	   echo
+#else
+	
+#	  fi
+#	fi
 
 #Обновляемся из salt
 salt1=$(systemctl is-active salt-minion >/dev/null 2>&1 && echo YES || echo NO)
@@ -114,5 +127,7 @@ fi
 chmod +rwxr+xr+x /usr/bin/wine
 
 
-env -i salt-call state.apply
+env -i salt-call state.apply | tail -n 7
+echo "$result_message1"
+echo "$result_message2"
